refactor(logging): extract //src/logging Bazel package
Promotes the generic levelled-logging primitive into its own Bazel
package per the Probe-8 layout (ADR-0003).

Renames:
  include/Logger.h  ->  src/logging/logger.h
  src/Logger.cpp    ->  src/logging/logger.cpp

New //src/logging:logging cc_library with features=[layering_check,
parse_headers]. Zero external deps. include_prefix="logging" stages
the header at the virtual path logging/logger.h so consumers whose
own include path does not have src/ on it (//shttps:shttps) resolve
the include without a workaround.

The visibility allowlist names //shttps:__pkg__ explicitly — logger
is a generic primitive (default namespace, no domain logic), shared
across modules as a support library rather than a SIPI-internal type.

//src:sipi_top adds //src/logging:logging as a transitive dep so
existing per-module consumers compile unchanged. //src:sipi_lib's
glob no longer needs to exclude Logger.cpp (package-boundary
semantics skip it now that src/logging/ has its own BUILD).
//include:headers loses the Logger.h export.
//shttps:shttps_headers's deps now reach logging directly instead of
//include:headers.

Call sites: 22 #include "Logger.h" rewritten to "logging/logger.h"
(the in-package logger.cpp uses "logger.h").

Resolves DEV-6487.
Resolves DEV-6488.
