+++ shttps/Server.cpp
@@ -31,7 +31,7 @@
 // #include "openssl/applink.c"
 
 
-#include "Logger.h"
+#include "logging/logger.h"
 #include "LuaServer.h"
 #include "Parsing.h"
 #include "Server.h"
